backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pgvector==0.3.2

# ─── Auth / Security ─────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9          # Rate limiting

# ─── Config ──────────────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ─── Task queue ──────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.8             # Includes redis.asyncio for pub/sub

# ─── AI — Claude (Anthropic) ─────────────────────────────────────────────────
anthropic==0.34.2

# ─── AI — Gemini (Google) ────────────────────────────────────────────────────
google-generativeai==0.8.3

# ─── LangGraph agents ────────────────────────────────────────────────────────
langgraph==0.2.28
langchain-anthropic==0.2.4
langchain-core==0.3.10

# ─── Scraping ────────────────────────────────────────────────────────────────
jobspy==1.1.80
playwright==1.47.0
playwright-stealth==1.0.6

# ─── ML / RL / Embeddings ────────────────────────────────────────────────────
sentence-transformers==3.1.1
scikit-learn==1.5.2      # cosine_similarity for reward model

# DPO training — optional, only needed when running rl/dpo_trainer.py
# Uncomment if you have a GPU and want to run DPO fine-tuning:
# trl==0.11.4
# transformers==4.45.2
# datasets==3.0.1
# torch==2.4.1           # Install separately with CUDA support if available

# ─── Utilities ───────────────────────────────────────────────────────────────
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1
